backend/requirements2.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiohttp-retry==2.9.1
aiosignal==1.4.0
asgiref==3.11.1
async-timeout==5.0.1
attrs==26.1.0
certifi==2026.2.25
charset-normalizer==3.4.7
dj-database-url==3.0.1
Django==4.2.29
django-cors-headers==4.9.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
frozenlist==1.8.0
gunicorn==23.0.0
idna==3.11
multidict==6.7.1
mysqlclient==2.2.7
packaging==26.0
pillow==11.3.0
propcache==0.4.1
psycopg2-binary==2.9.11
PyJWT==2.12.1
python-decouple==3.8
pytz==2026.1.post1
razorpay==2.0.1
requests==2.32.5
sqlparse==0.5.5
twilio==9.10.4
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
whitenoise==6.11.0
yarl==1.22.0
